+++ frontend/csi/config.go
@@ -1,8 +1,8 @@
-// Copyright 2018 NetApp, Inc. All Rights Reserved.
+// Copyright 2019 NetApp, Inc. All Rights Reserved.
 
 package csi
 
 const (
-	csiVersion    = "0.2"
+	csiVersion    = "1.0"
 	csiPluginName = "io.netapp.trident.csi"
 )